requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pymupdf>=1.24.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
google-generativeai>=0.8.3
requests>=2.31.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-magic>=0.4.27
slowapi>=0.1.9